backend/ai-core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
deepgram-sdk==3.2.7
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
openai==1.3.8
spacy==3.7.2
nltk==3.8.1
numpy==1.25.2
pandas==2.1.4